[fix] send osc

+++ src/track.cpp
@@ -99,10 +99,12 @@ void Track::send_osc(double beats) {
     p << osc::BeginMessage("/s_new")
         << "smooth" << Common::get().get_next_synth_id() << 0 << 0
         << "out" << this->bus_id
-        << "amp" << note.get_amp()
+        << "amp" << static_cast<float>(note.get_amp())
         << "sustain"
-        << this->scheduler->beats_to_seconds(note.get_duration())
-        << "freq" << Common::get().pitch_to_freq(note.get_pitch())
+        << static_cast<float>(
+            this->scheduler->beats_to_seconds(note.get_duration()))
+        << "freq" << static_cast<float>(
+            Common::get().pitch_to_freq(note.get_pitch()))
       << osc::EndMessage;
     Common::get().get_transmit_socket()->Send(p.Data(), p.Size());
